% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_classes.R
\name{get_col_classes}
\alias{get_col_classes}
\title{get_col_classes}
\usage{
get_col_classes(data, view = TRUE)
}
\arguments{
\item{data}{Dataframe}

\item{view}{Whether to automatically view results in RStudio}
}
\value{
Janes equipment data.
}
\description{
Returns View() with all column names and classes
}
