% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_sizes.R
\name{weigh_cols}
\alias{weigh_cols}
\title{get_col_sizes}
\usage{
weigh_cols(data, view = TRUE)
}
\arguments{
\item{data}{Dataframe}
}
\value{
Sizes of columns in DF as MBs
}
\description{
Returns column sizes as MBs
}
