% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_environment.R
\name{weigh_environment}
\alias{weigh_environment}
\title{weigh_environment}
\usage{
weigh_environment(env)
}
\value{
Environment object sizes
}
\description{
Returns size of each object in environment
}
