% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_janes.R
\name{write_janes}
\alias{write_janes}
\title{write_janes}
\usage{
write_janes(data, file_name, file_type = "csv")
}
\arguments{
\item{data}{data}

\item{file_name}{Name of file}

\item{file_type}{CSV or XLSX}
}
\value{
Writes files to
}
\description{
Writes CSVs
}
